refactor(paths): extract PathTypeResolver from PathExpressionDecomposer
